# docs/swagger/swagger.yaml
definitions:
  AddEmployeeInput:
    properties:
      employment_type:
        enum:
        - full_time
        - part_time
        - contract
        - intern
        - temporary
        - consultant
        type: string
      job_title:
        maxLength: 100
        minLength: 2
        type: string
      membership_status:
        enum:
        - active
        - on_leave
        - suspended
        - resigned
        - terminated
        type: string
      phone:
        type: string
      role_id:
        type: string
    required:
    - employment_type
    - job_title
    - membership_status
    - phone
    - role_id
    type: object
  AddEmployeeOutput:
    properties:
      id:
        type: string
    type: object
...
  CreateRoleInput:
    properties:
      permissions:
        items:
          type: string
        minItems: 1
        type: array
      title:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - permissions
    - title
    type: object
  CreateRoleOutput:
    properties:
      id:
        type: string
    type: object
...
  EmployeeOutput:
    properties:
      employment_type:
        type: string
      first_name:
        type: string
      id:
        type: string
      job_title:
        type: string
      last_name:
        type: string
      membership_status:
        type: string
      phone:
        type: string
      role_id:
        type: string
      user_id:
        type: string
    type: object
  Error:
    properties:
      code:
        example: INVALID_ARGUMENT
        type: string
      message:
        example: invalid request body
        type: string
      meta:
        additionalProperties:
          type: string
        example:
          error: invalid_length
          field: username
        type: object
    type: object
...
  LoginInput:
    properties:
      password:
        maxLength: 72
        minLength: 8
        type: string
      phone:
        type: string
    required:
    - password
    - phone
    type: object
  LoginOutput:
    properties:
      access_token:
        type: string
    type: object
...
  RoleOutput:
    properties:
      id:
        type: string
      is_owner:
        type: boolean
      permissions:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
...
  UpdateRoleInput:
    properties:
      permissions:
        items:
          type: string
        minItems: 1
        type: array
      title:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - permissions
    - title
    type: object
  VerifyOTPAndRegisterInput:
    properties:
      code:
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
      phone:
        type: string
    required:
    - code
    - password
    - phone
    type: object
  VerifyOTPAndRegisterOutput:
    properties:
      access_token:
        type: string
    type: object
info:
  contact:
    email: support@trueflow.ir
    name: API Support
    url: https://trueflow.ir/support
  description: TrueFlow is a modern human resource management (HRM) platform designed
    to integrate organizational processes, manage teams, control access, and increase
    organizational productivity.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: TrueFlow Api Document
paths:
  /authentication/login:
    post:
      consumes:
      - application/json
      description: Logins the user and returns an access token.
      parameters:
      - description: Register request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LoginOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      summary: Login user
      tags:
      - Authentication
...
  /authentication/verify-otp:
    post:
      consumes:
      - application/json
      description: Verifies OTP, registers the user and returns an access token.
      parameters:
      - description: Register request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/VerifyOTPAndRegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/VerifyOTPAndRegisterOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      summary: Register user
      tags:
      - Authentication
...
  /organization/{organization_id}/employee:
    get:
      parameters:
      - description: Organization ID
        in: path
        name: organization_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/EmployeeOutput'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - BearerAuth: []
      summary: List employees
      tags:
      - Employee
    post:
      consumes:
      - application/json
      parameters:
      - description: Organization ID
        in: path
        name: organization_id
        required: true
        type: string
      - description: Add employee request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/AddEmployeeInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/AddEmployeeOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - BearerAuth: []
      summary: Add employee
      tags:
      - Employee
...
  /organization/{organization_id}/role:
    get:
      description: Returns every role defined for an organization.
      parameters:
      - description: Organization ID
        in: path
        name: organization_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/RoleOutput'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - BearerAuth: []
      summary: List roles
      tags:
      - Role
    post:
      consumes:
      - application/json
      description: Creates a new role for an organization.
      parameters:
      - description: Organization ID
        in: path
        name: organization_id
        required: true
        type: string
      - description: Create role request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreateRoleInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CreateRoleOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - BearerAuth: []
      summary: Create role
      tags:
      - Role
  /organization/{organization_id}/role/{role_id}:
    delete:
      parameters:
      - description: Organization ID
        in: path
        name: organization_id
        required: true
        type: string
      - description: Role ID
        in: path
        name: role_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - BearerAuth: []
      summary: Delete role
      tags:
      - Role
    patch:
      consumes:
      - application/json
      parameters:
      - description: Organization ID
        in: path
        name: organization_id
        required: true
        type: string
      - description: Role ID
        in: path
        name: role_id
        required: true
        type: string
      - description: Update role request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UpdateRoleInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - BearerAuth: []
      summary: Update role
      tags:
      - Role
...
securityDefinitions:
  BearerAuth:
    description: Enter the JWT token with the `Bearer ` prefix
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"